Calculate the Depth of Latest Common Ancestor (LCA) for Cell Lineage Tree Nodes.

This function calculates the depth of the latest common ancestor (LCA) for pairs of leaf nodes in a cell lineage tree. It computes the distance from each node to the root (distance_to_root), the distance to the farthest leaf (distance_to_farthest_leaf), and normalized versions of these quantities. It then enumerates all pairs of leaf nodes, finds their LCA, and reports the depth/height metrics of that LCA.

Usage

calculate_lca_depths(tree.path, mc.cores = parallel::detectCores())

Arguments

tree.path

The input path of the cell lineage tree in Newick format (for example, a path obtained by system.file("extdata", "HSC.nwk", package = "PASTRI")).

mc.cores

Integer. Desired number of parallel workers for internal computations. On Windows, values greater than 1 are not supported by parallel::mclapply and will be silently coerced to 1.

Value

A data frame with the following columns:

node1: The label of the first leaf node in the pair.

node2: The label of the second leaf node in the pair.

subtree_root: The internal node ID corresponding to the latest common ancestor (LCA).

lca_depth: The distance from the LCA to the farthest leaf in that subtree.

lca_depth_norm: The normalized distance of the LCA to the farthest leaf.

lca_normalized_height: The normalized height (1 - normalized depth) of the LCA.

Examples

if (FALSE) { # \dontrun{
tree_file <- system.file("extdata", "HSC.nwk", package = "PASTRI")
res <- calculate_lca_depths(tree.path = tree_file, mc.cores = 2)
head(res)
} # }