Calculate Transition Matrix from Two-Cell Correlation

This function calculates cell transition matrices based on two-cell correlation functions C(u). It first computes the frequency of each cell type (using only single-leaf nodes), then builds a standardized cell type combination to avoid duplicate counts. Using the node pair depth data and selected depth measure (e.g., "lca_normalized_height" or "lca_depth"), the function computes the correlation frequencies for each depth. For each depth (up to a maximum threshold fi_depth), a corresponding transition matrix is computed using eigenvalue decomposition, where the eigenvalues are adjusted according to the depth. Finally, the average transition matrix from all depths is used as the initial guess for an optimization procedure (via the L-BFGS-B method) to derive the optimal transition matrix.

Usage

calculate_transition_matrix(
  node_pair_depth,
  node_info,
  sel_u,
  use_cores,
  fi_depth,
  bound_matrix,
  cell_type_list
)

Arguments

node_pair_depth

A data frame containing node pair depth information and related metrics (e.g., normalized LCA height).

node_info

A data frame containing detailed information for each node, including columns 'celltype' (cell type) and 'nodeLabel' (node label).

sel_u

A string indicating the depth column to be used (e.g., "lca_normalized_height" or "lca_depth").

use_cores

An integer specifying the number of cores to use for parallel computation (default is 80).

fi_depth

A numeric value indicating the maximum depth threshold for computing the transition matrix.

bound_matrix

A constrained matrix

cell_type_list

A list of selected cell types

Value

A list containing:

all_trans_matrix_list: A list of transition matrices computed for different depths.

optimal_trans_matrix: The optimal transition matrix obtained from the optimization procedure.

Examples

# Suppose you have node pair depth data (node_pair_depth) and node information (node_info), and you select
# "lca_normalized_height" as the depth measure with fi_depth set to 0.2:
result <- calculate_transition_matrix(node_pair_depth, node_info, sel_u = "lca_normalized_height", use_cores = 80, fi_depth = 0.2, bound_matrix, cell_type_list)
#> Error: object 'node_info' not found
# To view the optimal transition matrix:
print(result$optimal_trans_matrix)
#> Error: object 'result' not found