Calculate Transition Matrix Error

This function computes the total error between an initial guess transition matrix and a list of observed transition matrices. The error is defined as the cumulative Frobenius norm difference between the guess and each observed matrix.

Usage

calculate_transition_matrix_error(
  initial_vector,
  observed_matrices,
  matrix_size,
  type_list
)

Arguments

initial_vector

A numeric vector representing the flattened initial guess for the transition matrix.

observed_matrices

A list of observed transition matrices.

matrix_size

The size (number of rows/columns) of the transition matrix.

type_list

A character vector of cell types used to set the row and column names of the matrix.

Value

A numeric value representing the total error.

Examples

# For example, if you have a flattened 5x5 initial guess vector and two 5x5 observed matrices:
initial_vec <- runif(25)  # generate a random initial guess for a 5x5 matrix
observed_list <- list(matrix(runif(25), nrow = 5), matrix(runif(25), nrow = 5))
cell_types <- c("C4", "C5", "C6", "C7", "C8")
error_value <- calculate_transition_matrix_error(initial_vec, observed_list, 5, cell_types)